Integrate vlc playback with motion control
